+++ Godeps/_workspace/src/gopkg.in/bsm/ratelimit.v1/ratelimit.go
@@ -25,7 +25,7 @@ import (
 
 // RateLimit instances are thread-safe.
 type RateLimiter struct {
-	allowance, max, unit, lastCheck uint64
+	rate, allowance, max, unit, lastCheck uint64
 }
 
 // New creates a new rate limiter instance
@@ -39,22 +39,23 @@ func New(rate int, per time.Duration) *RateLimiter {
 	}
 
 	return &RateLimiter{
-		allowance: uint64(rate) * nano, // store our allowance, in ns units
+		rate:      uint64(rate),        // store the rate
+		allowance: uint64(rate) * nano, // set our allowance to max in the beginning
 		max:       uint64(rate) * nano, // remember our maximum allowance
 		unit:      nano,                // remember our unit size
 
-		lastCheck: uint64(time.Now().UnixNano()),
+		lastCheck: unixNano(),
 	}
 }
 
 // Limit returns true if rate was exceeded
 func (rl *RateLimiter) Limit() bool {
 	// Calculate the number of ns that have passed since our last call
-	now := uint64(time.Now().UnixNano())
+	now := unixNano()
 	passed := now - atomic.SwapUint64(&rl.lastCheck, now)
 
 	// Add them to our allowance
-	current := atomic.AddUint64(&rl.allowance, passed)
+	current := atomic.AddUint64(&rl.allowance, passed*rl.rate)
 
 	// Ensure our allowance is not over maximum
 	if current > rl.max {
@@ -80,4 +81,9 @@ func (rl *RateLimiter) Undo() {
 	if current > rl.max {
 		atomic.AddUint64(&rl.allowance, rl.max-current)
 	}
+}
+
+// now as unix nanoseconds
+func unixNano() uint64 {
+	return uint64(time.Now().UnixNano())
 }