+++ Godeps/_workspace/src/gopkg.in/redis.v3/pool.go
@@ -20,7 +20,7 @@ type pool interface {
 	First() *conn
 	Get() (*conn, bool, error)
 	Put(*conn) error
-	Remove(*conn) error
+	Remove(*conn, error) error
 	Len() int
 	FreeLen() int
 	Close() error
@@ -130,14 +130,14 @@ type connPool struct {
 
 	_closed int32
 
-	lastDialErr error
+	lastErr atomic.Value
 }
 
 func newConnPool(opt *Options) *connPool {
 	p := &connPool{
 		dialer: newConnDialer(opt),
 
-		rl:        ratelimit.New(2*opt.getPoolSize(), time.Second),
+		rl:        ratelimit.New(3*opt.getPoolSize(), time.Second),
 		opt:       opt,
 		conns:     newConnList(opt.getPoolSize()),
 		freeConns: make(chan *conn, opt.getPoolSize()),
@@ -163,8 +163,12 @@ func (p *connPool) First() *conn {
 		select {
 		case cn := <-p.freeConns:
 			if p.isIdle(cn) {
-				p.conns.Remove(cn)
-				continue
+				var err error
+				cn, err = p.replace(cn)
+				if err != nil {
+					log.Printf("redis: replace failed: %s", err)
+					continue
+				}
 			}
 			return cn
 		default:
@@ -181,8 +185,12 @@ func (p *connPool) wait() *conn {
 		select {
 		case cn := <-p.freeConns:
 			if p.isIdle(cn) {
-				p.Remove(cn)
-				continue
+				var err error
+				cn, err = p.replace(cn)
+				if err != nil {
+					log.Printf("redis: replace failed: %s", err)
+					continue
+				}
 			}
 			return cn
 		case <-deadline:
@@ -196,15 +204,15 @@ func (p *connPool) wait() *conn {
 func (p *connPool) new() (*conn, error) {
 	if p.rl.Limit() {
 		err := fmt.Errorf(
-			"redis: you open connections too fast (last error: %v)",
-			p.lastDialErr,
+			"redis: you open connections too fast (last_error=%q)",
+			p.loadLastErr(),
 		)
 		return nil, err
 	}
 
 	cn, err := p.dialer()
 	if err != nil {
-		p.lastDialErr = err
+		p.storeLastErr(err.Error())
 		return nil, err
 	}
 
@@ -247,8 +255,9 @@ func (p *connPool) Get() (cn *conn, isNew bool, err error) {
 func (p *connPool) Put(cn *conn) error {
 	if cn.rd.Buffered() != 0 {
 		b, _ := cn.rd.Peek(cn.rd.Buffered())
-		log.Printf("redis: connection has unread data: %q", b)
-		return p.Remove(cn)
+		err := fmt.Errorf("redis: connection has unread data: %q", b)
+		log.Print(err)
+		return p.Remove(cn, err)
 	}
 	if p.opt.getIdleTimeout() > 0 {
 		cn.usedAt = time.Now()
@@ -257,16 +266,26 @@ func (p *connPool) Put(cn *conn) error {
 	return nil
 }
 
-func (p *connPool) Remove(cn *conn) error {
-	// Replace existing connection with new one and unblock waiter.
+func (p *connPool) replace(cn *conn) (*conn, error) {
 	newcn, err := p.new()
 	if err != nil {
-		log.Printf("redis: new failed: %s", err)
-		return p.conns.Remove(cn)
+		_ = p.conns.Remove(cn)
+		return nil, err
 	}
-	err = p.conns.Replace(cn, newcn)
+	_ = p.conns.Replace(cn, newcn)
+	return newcn, nil
+}
+
+func (p *connPool) Remove(cn *conn, reason error) error {
+	p.storeLastErr(reason.Error())
+
+	// Replace existing connection with new one and unblock waiter.
+	newcn, err := p.replace(cn)
+	if err != nil {
+		return err
+	}
 	p.freeConns <- newcn
-	return err
+	return nil
 }
 
 // Len returns total number of connections.
@@ -314,6 +333,17 @@ func (p *connPool) reaper() {
 	}
 }
 
+func (p *connPool) storeLastErr(err string) {
+	p.lastErr.Store(err)
+}
+
+func (p *connPool) loadLastErr() string {
+	if v := p.lastErr.Load(); v != nil {
+		return v.(string)
+	}
+	return ""
+}
+
 //------------------------------------------------------------------------------
 
 type singleConnPool struct {
@@ -341,7 +371,7 @@ func (p *singleConnPool) Put(cn *conn) error {
 	return nil
 }
 
-func (p *singleConnPool) Remove(cn *conn) error {
+func (p *singleConnPool) Remove(cn *conn, _ error) error {
 	if p.cn != cn {
 		panic("p.cn != cn")
 	}
@@ -415,38 +445,34 @@ func (p *stickyConnPool) put() (err error) {
 func (p *stickyConnPool) Put(cn *conn) error {
 	defer p.mx.Unlock()
 	p.mx.Lock()
-	if p.cn != cn {
-		panic("p.cn != cn")
-	}
 	if p.closed {
 		return errClosed
+	}
+	if p.cn != cn {
+		panic("p.cn != cn")
 	}
 	return nil
 }
 
-func (p *stickyConnPool) remove() (err error) {
-	err = p.pool.Remove(p.cn)
+func (p *stickyConnPool) remove(reason error) (err error) {
+	err = p.pool.Remove(p.cn, reason)
 	p.cn = nil
 	return err
 }
 
-func (p *stickyConnPool) Remove(cn *conn) error {
+func (p *stickyConnPool) Remove(cn *conn, _ error) error {
 	defer p.mx.Unlock()
 	p.mx.Lock()
+	if p.closed {
+		return errClosed
+	}
 	if p.cn == nil {
 		panic("p.cn == nil")
 	}
 	if cn != nil && p.cn != cn {
 		panic("p.cn != cn")
 	}
-	if p.closed {
-		return errClosed
-	}
-	if cn == nil {
-		return p.remove()
-	} else {
-		return nil
-	}
+	return nil
 }
 
 func (p *stickyConnPool) Len() int {
@@ -467,6 +493,15 @@ func (p *stickyConnPool) FreeLen() int {
 	return 0
 }
 
+func (p *stickyConnPool) Reset(reason error) (err error) {
+	p.mx.Lock()
+	if p.cn != nil {
+		err = p.remove(reason)
+	}
+	p.mx.Unlock()
+	return err
+}
+
 func (p *stickyConnPool) Close() error {
 	defer p.mx.Unlock()
 	p.mx.Lock()
@@ -479,7 +514,8 @@ func (p *stickyConnPool) Close() error {
 		if p.reusable {
 			err = p.put()
 		} else {
-			err = p.remove()
+			reason := errors.New("redis: sticky not reusable connection")
+			err = p.remove(reason)
 		}
 	}
 	return err