+++ CHANGELOG.md
@@ -1,1 +1,15 @@
 # CHANGELOG
+
+## 0.5.0
+
+Dockyard 0.5.0 is an incremental release with new features,bug fixes and new support for Go 1.5.3.
+
+#### New features
+
+- New support for ACI storage,support rkt fetch&push ACIs from Dockyard,adapt to rkt and acpush client.
+- New support for OSS that is a distributed replicated object store.See [OSS](oss/README.md) documentation.
+
+#### Bug fixes
+
+- Fix and update the system test case.
+- Fix V2 protocol bug.